kefir_api/permissions.py
from rest_framework import permissions


class IsAdminOrAuthorOrReadOnly(permissions.BasePermission):
    def has_permission(self, request, view):
        return request.user.is_authenticated

    def has_object_permission(self, request, view, obj):
        return(
                request.user == obj or
                request.user.is_superuser and
                request.user.is_authenticated
        )
